pyproject.toml
[project]
name = "arxivsmart"
version = "0.0.1"
description = "arXiv API proxy with rate limiting and MCP server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
    "defusedxml>=0.7.1",
    "markdownify>=0.14.1",

    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-randomly>=3.15.0",
    "pytest-timeout>=2.2.0",

    # Linting & Formatting
    "ruff>=0.1.0",

    # Type Checking
    "mypy>=1.7.0",
    "pyright>=1.1.0",
    "types-PyYAML>=6.0.12",

    # Security & Quality
    "bandit[toml]>=1.7.8",
    "deadcode>=2.4.0",
    "deptry>=0.20.0",
    "pygount>=1.6.4",
    "codespell>=2.3.0",
    "pip-audit>=2.7.0",
    "semgrep>=1.144.0",

    # Git Hooks
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arxivsmart"]

[tool.ruff]
line-length = 140
indent-width = 4
target-version = "py312"
exclude = [".git", ".ruff_cache", ".venv", ".vscode", "reports", "stubs"]

[tool.ruff.lint]
# Core errors & warnings, Google-style docs, types, bugbear, quotes, imports, naming
# Plus: complexity, modern syntax, performance, simplifications, tidy imports
select = [
  "E", "F",   # pycodestyle errors, pyflakes
  "D",        # pydocstyle (Google-style)
  "ANN",      # flake8-annotations
  "B",        # flake8-bugbear
  "Q",        # flake8-quotes
  "I",        # isort
  "N",        # pep8-naming
  "C90",      # mccabe complexity
  "UP",       # pyupgrade (modern Python 3.12+ syntax)
  "PERF",     # perflint (performance anti-patterns)
  "SIM",      # flake8-simplify
  "TID",      # flake8-tidy-imports
]
ignore = [
  "D107",    # don't require __init__ docstring (class docstring is fine)
  "N802",    # allow non-lowercase function names (if needed)
]

[tool.ruff.lint.mccabe]
# Flag functions with complexity > 10
max-complexity = 10

[tool.ruff.lint.flake8-tidy-imports]
# Ban relative imports for cleaner, more explicit code
ban-relative-imports = "all"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D", "ANN"]
"tests_e2e/**/*.py" = ["D", "ANN"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [".", "src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# pytest-timeout: prevent tests from hanging indefinitely
timeout = 30
timeout_method = "thread"
# pytest-randomly: randomize test order to catch hidden dependencies

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true

# Limit what mypy considers part of the project
files = ["src", "tests"]

# Custom type stubs for third-party libraries
mypy_path = ["stubs"]

# CI ergonomics
show_error_codes = true
pretty = true

# Exclude virtual environment
exclude = [".venv/"]


# Tests – still type checked, but not fully strict
[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
check_untyped_defs = false

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]

[tool.coverage.html]
directory = "reports/coverage/html"

[tool.coverage.xml]
output = "reports/coverage/coverage.xml"

[tool.bandit]
targets = ["src"]
exclude = [
    "tests",
    ".venv",
    "venv",
]
# Reasonable default for CI
severity = "MEDIUM"
confidence = "MEDIUM"

[tool.deptry]
# Source scanning - only check src/ directory
extend_exclude = [
    "build",
    "dist",
    "docs",
    "scripts",
    "reports",
    "notes",
]

# Ignore Jupyter notebooks (not part of production code)
ignore_notebooks = true

# First-party modules - tell deptry what belongs to this project
known_first_party = [
    "src",
]

# All dependencies are in the main group
pep621_dev_dependency_groups = []

# Keep all rules enabled (DEP001-DEP005)
ignore = []

# CI/Reporting integration
json_output = "reports/deptry/deptry.json"

[tool.deptry.per_rule_ignores]
# Runtime libraries detected via dynamic imports or framework conventions
DEP002 = [
    "fastapi",
    "uvicorn",
    "httpx",
    "defusedxml",
    "markdownify",
    # CLI dev tools — invoked by justfile, never imported in source
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-randomly",
    "pytest-timeout",
    "ruff",
    "mypy",
    "pyright",
    "types-PyYAML",
    "bandit",
    "deptry",
    "pygount",
    "codespell",
    "deadcode",
    "pip-audit",
    "semgrep",
    "pre-commit",
]

[tool.codespell]
# Paths to check
check-filenames = true
check-hidden = false

# Directories to skip
skip = ".git,.venv,venv,build,dist,reports,*.lock,uv.lock"

# Use external ignore file for maintainability
ignore-words = "config/codespell/ignore.txt"

[tool.deadcode]
exclude = [
    ".venv",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    "stubs",
    "reports",
    "config",
    "prompts",
    "docs",
    "tests",
    "tests_e2e",
]

# Abstract methods, pytest fixtures, lifespan managers
ignore-definitions-if-decorated-with = [
    "abstractmethod",
    "pytest.fixture",
    "asynccontextmanager",
]

# Pydantic field validators (called by metaclass) and model_config attributes
ignore-names = [
    "validate_*",
    "model_config",
]

# Pydantic model fields (used via serialization) and ABC bodies (used via subclasses)
ignore-bodies-if-inherits-from = [
    "BaseModel",
    "ABC",
]

# FastAPI route files — functions are registered by decorators, not called directly
# Client classes — library API consumed by external callers / MCP server
# main.py — uvicorn reload-mode factory function
ignore-names-in-files = [
    "src/arxivsmart/api/routes_*.py",
    "src/arxivsmart/clients/*.py",
    "src/main.py",
]
